requirements.txt
Django==4.0.3
djangorestframework==3.13.1
python-decouple==3.6
django-extensions==3.1.5